+++ Sources/MilitarySymbologyKit/Extensions/MilitarySymbol+Ext.swift
@@ -6,8 +6,30 @@ import MilitarySymbologyAssets
 import OSLog
 import SwiftUI
 
+// MARK: - Conformance to protocols
+extension MilitarySymbol: Comparable {
+    public static func < (lhs: MilitarySymbol, rhs: MilitarySymbol) -> Bool {
+        if lhs.dimention.id == rhs.dimention.id {
+            if lhs.entity.id == rhs.entity.id {
+                if lhs.entityType.id == rhs.entityType.id {
+                    return lhs.entitySubtype.id > rhs.entitySubtype.id
+                }
+                
+                return lhs.entity.id < rhs.entity.id
+            }
+            
+            return lhs.entity.id < rhs.entity.id
+        }
+        
+        return lhs.dimention.id < rhs.dimention.id
+    }
+}
+
+// MARK: - initializers
+
 public extension MilitarySymbol {
     /// Init with *default* values.
+    /// Possible to change Status Amplifiers style.
     init(isAlternateStatusAmplifiers: Bool = true) {
         context = .reality
         standardIdentity = .pending
@@ -22,12 +44,112 @@ public extension MilitarySymbol {
         isCivilian = false
         self.isAlternateStatusAmplifiers = isAlternateStatusAmplifiers
     }
+    
+    /// Init from *SIDC* code.
+    /// Possible to change Status Amplifiers style.
+    init(sidc: String, isAlternateStatusAmplifiers: Bool = false) throws {
+        if sidc.count != 20 {
+            throw MilitarySymbolError.sidcIsNot20
+        } else {
+            guard let context = Context(rawValue: sidc[2]) else {
+                print("Context")
+                throw MilitarySymbolError.contextParcingFailed
+            }
+            self.context = context
+            
+            guard let standardIdentity = StandardIdentity(rawValue: sidc[3]) else {
+                print("StandardIdentity")
+                throw MilitarySymbolError.standardIdentityParcingFailed
+            }
+            self.standardIdentity = standardIdentity
+            
+            guard let dimention = Dimension(rawValue: sidc[4] + sidc[5]) else {
+                print("Dimension")
+                throw MilitarySymbolError.dimentionParcingFailed
+            }
+            self.dimention = dimention
+            
+            guard let status = Status(rawValue: sidc[6]) else {
+                print("Status")
+                throw MilitarySymbolError.statusParcingFailed
+            }
+            self.status = status
+            
+            guard let hqtfd = HQTFD(rawValue: sidc[7]) else {
+                print("HQTFD")
+                throw MilitarySymbolError.hqtfdParcingFailed
+            }
+            self.hqtfd = hqtfd
+            
+            guard let amplifier = Amplifier(rawValue: sidc[8]) else {
+                print("Amplifier")
+                throw MilitarySymbolError.amplifierParcingFailed
+            }
+            self.amplifier = amplifier
+            
+            guard let descriptor = amplifier.descriptors.first(where: { $0.id == sidc[9] }) else {
+                print("Descriptor")
+                throw MilitarySymbolError.descriptorParcingFailed
+            }
+            self.descriptor = descriptor
+            
+            let entityDigits: String = sidc[10] + sidc[11]
+            guard let entity = dimention.entities.first(where: { $0.id == entityDigits }) else {
+                print("Entity")
+                throw MilitarySymbolError.entityParcingFailed
+            }
+            self.entity = entity
+            
+            let entityTypeDigits: String = sidc[12] + sidc[13]
+            guard let entityType = entity.types.first(where: { $0.id == entityTypeDigits }) else {
+                print("EntityType")
+                throw MilitarySymbolError.entityTypeParcingFailed
+            }
+            self.entityType = entityType
+            
+            let entitySybTypeDigits = sidc[14] + sidc[15]
+            guard let entitySubtype = entityType.subtypes.first(where: { $0.id == entitySybTypeDigits }) else {
+                print("EntitySubtype")
+                throw MilitarySymbolError.entitySubtypeParcingFailed
+            }
+            self.entitySubtype = entitySubtype
+            
+            self.isAlternateStatusAmplifiers = isAlternateStatusAmplifiers
+        }
+    }
+    
+    /// Init for custom values.
+    init(
+        context: Context,
+        standardIdentity: StandardIdentity,
+        dimention: Dimension,
+        status: Status,
+        hqtfd: HQTFD,
+        amplifier: Amplifier,
+        descriptor: AnyDescriptor,
+        entity: AnyEntity,
+        entityType: AnyEntityType,
+        entitySubtype: AnyEntitySubtype,
+        isAlternateStatusAmplifiers: Bool = true
+    ) {
+        self.context = context
+        self.standardIdentity = standardIdentity
+        self.dimention = dimention
+        self.status = status
+        self.hqtfd = hqtfd
+        self.amplifier = amplifier
+        self.descriptor = descriptor
+        self.entity = entity
+        self.entityType = entityType
+        self.entitySubtype = entitySubtype
+        self.isCivilian = false
+        self.isAlternateStatusAmplifiers = isAlternateStatusAmplifiers
+    }
 }
 
+// MARK: - Asset name generators
 public extension MilitarySymbol {
     // MARK: - FRAME
-
-    // All good.
 
     /// **Uses SIDC positions 3-7**, with an underscore between the first digit in the name and the last digit in the name. Purple filled frames for Civilian units, equipment, and installations have a ‘c’ at the end of the file name.
     var frameAssetName: String? {
@@ -94,8 +216,6 @@ public extension MilitarySymbol {
 
     // MARK: - AMPLIFIER
 
-    // All good.
-
     /// Uses SIDC positions 4 and 9-10.
     var amplifierAssetName: String? {
         switch amplifier {
@@ -114,6 +234,8 @@ public extension MilitarySymbol {
         }
     }
 
+    // MARK: - HQTFD
+    
     /// Uses SIDC **positions 4-6 (StandardIdentity-Dimention)** and position **8 (hqtfd)**.
     var hqtfdAssetName: String? {
         if hqtfd == .none {
@@ -142,6 +264,8 @@ public extension MilitarySymbol {
             return result
         }
     }
+    
+    // MARK: - OCA
 
     var ocaAssetName: String? {
         if isAlternateStatusAmplifiers {
@@ -165,6 +289,8 @@ public extension MilitarySymbol {
             }
         }
     }
+    
+    // MARK: - Main icons
 
     /// Uses SIDC positions **5-6 (Symbol Set) and 11-16 (Entity, EntityType, EntitySubtype)**.
     var mainIconAssetName: String? {
@@ -313,7 +439,12 @@ public extension MilitarySymbol {
             return result
         }
     }
+}
 
+// MARK: - Views
+
+public extension MilitarySymbol {
+    
     func makeView(size: CGFloat? = nil) -> some View {
         ZStack {
             
@@ -353,9 +484,7 @@ public extension MilitarySymbol {
         }
         .frame(width: size, height: size)
     }
-}
-
-public extension MilitarySymbol {
+    
     static func makeUnknownSymbolView(size: CGFloat? = nil) -> some View {
         Image(systemName: "questionmark.diamond.fill")
             .resizable()
@@ -363,5 +492,58 @@ public extension MilitarySymbol {
             .frame(width: size)
             .symbolRenderingMode(.hierarchical)
             .foregroundStyle(.orange)
+    }
+}
+
+// MARK: - Filtering
+// TODO: Delete this later.
+
+public extension MilitarySymbol {
+    
+    static func searched(text: String, currentSymbol: MilitarySymbol?) -> [MilitarySymbol] {
+        if text.isEmpty {
+            return []
+        } else {
+            var result: [MilitarySymbol] = []
+            Dimension.allCases.forEach { dimention in
+                dimention.entities.forEach { entity in
+                    entity.types.forEach { entityType in
+                        entityType.subtypes.forEach { entitySubtype in
+                            if entity.name.localizedStandardContains(text)
+                                || entityType.name.localizedStandardContains(text)
+                                || entitySubtype.name.localizedStandardContains(text)
+                            {
+                                if let currentSymbol {
+                                    result.append(
+                                        /// Version with current symbol properties.
+                                        MilitarySymbol(context: currentSymbol.context,
+                                                       standardIdentity: currentSymbol.standardIdentity,
+                                                       dimention: dimention,
+                                                       status: currentSymbol.status,
+                                                       hqtfd: currentSymbol.hqtfd,
+                                                       amplifier: currentSymbol.amplifier,
+                                                       descriptor: currentSymbol.descriptor,
+                                                       entity: entity,
+                                                       entityType: entityType,
+                                                       entitySubtype: entitySubtype,
+                                                       isCivilian: currentSymbol.isCivilian,
+                                                       isAlternateStatusAmplifiers: currentSymbol.isAlternateStatusAmplifiers)
+                                    )
+                                } else {
+                                    result.append(
+                                        /// Version with default symbol properties.
+                                        MilitarySymbol(dimention: dimention,
+                                                       entity: entity,
+                                                       entityType: entityType,
+                                                       entitySubtype: entitySubtype)
+                                    )
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            return result
+        }
     }
 }
